requirements.txt
Django==4.2.3
djangorestframework==3.14.0
python-decouple==3.8
django-cors-headers==4.1.0
django-filter==23.2
psycopg2-binary==2.9.6